backend/pyproject.toml
[tool.poetry]
name = "orgbook-publisher"
version = "0.1.0"
description = ""
authors = ["PatStLouis <patrick.st-louis@opsecid.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
pydantic-settings = "^2.5.2"
validators = "^0.34.0"
aries-askar = "^0.3.2"
multiformats = "^0.3.1.post4"
canonicaljson = "^2.0.0"
bitstring = "^4.2.3"
bs4 = "^0.0.2"
jsonpath-ng = "^1.7.0"
jinja2 = "^3.1.4"
httpx = "^0.27.2"
segno = "^1.6.1"
ruff = "^0.7.0"
pystache = "^0.6.5"
pyjwt = "^2.9.0"
qrcode = "^8.0"
pillow = "^11.0.0"
python-multipart = "^0.0.17"
blake3 = "^0.4.1"
pymongo = "^4.10.1"
base58 = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
